fix: resolve remaining controllers/api compile errors
Four more independent, pre-existing bugs in the same "never actually
built" category as commit 851552c:

- controllers/api/analytics.go: csv.Writer.Flush() returns no value in
  the standard library; the error (if any) is checked via Writer.Error()
  afterward.
- auth/jwt.go: Claims.UserID and GenerateJWT's parameter were uint,
  while every other user-ID type in this codebase (models.User.Id,
  the ctx "user_id" value read back in controllers/api/auth.go's
  profile handler) is int64. Changed to int64 for consistency instead
  of casting at each call site.
- controllers/api/auth.go: undefined models.NowUTC (never existed
  anywhere in the codebase) — replaced with time.Now().UTC().
- controllers/api/campaign.go: campaignCreateRequest embeds
  DeliveryConfig rather than promoting its fields, so
  req.DelayBetweenMs etc. needed to be req.DeliveryConfig.X (also
  correcting DelayBetweenMs to the field's real name,
  DelayBetweenEmails). Separately, storeCampaignSMTPs built a
  CampaignSMTP struct literal using CampaignId/SMTPId/Priority, none of
  which exist on that type (real fields: CampaignID/SMTPID; there is no
  Priority concept in the model at all, and nothing reads it).

+++ auth/jwt.go
@@ -41,13 +41,13 @@ func secretOnceInit() {
 }
 
 type Claims struct {
-	UserID   uint   `json:"user_id"`
+	UserID   int64  `json:"user_id"`
 	Username string `json:"username"`
 	Role     string `json:"role"`
 	jwt.RegisteredClaims
 }
 
-func GenerateJWT(userID uint, username, role string) (string, error) {
+func GenerateJWT(userID int64, username, role string) (string, error) {
 	secretOnceInit()
 	claims := &Claims{
 		UserID:   userID,
